Support version compatibility check in config (#1505)
This adds a new key, `tool-version,` to express version requirements. For example,

```toml
tool-version = "=0.11.1"
```

will enforce that the Exograph cli version must be exactly "0.11.1".

We support all the standard server requirement expressions, such as `0.1.1`, `=0.1.1`, `1`, `>= 1`, and so on. For more information, see the docs for the `server` crate (https://docs.rs/semver/1.0.25/semver).