Introduce `exo update` command (#1643)
Updates Exograph to the latest version.

Also, check if a new version is available in any `exo` command (except `exo update`) to notify availability of a newer version.

```sh
$ exo build
A new version (0.15.3) of Exograph is available. Run `exo update` to update.
```

When offline or upon hitting Github's rate limiting, it silently skips the update checks to avoid breaking the execution of the main command such as `exo dev` or `exo build`.