Simplify Fly.io deployment (#1655)
Previously, we had a complex setup where we created two Dockerfiles (one for building and another for running). We also ran migrations during the build phase. With recent changes to Fly.io (specifically the use of Depot for building), this setup no longer works.

Now, we produce only one Dockerfile and shift the migration responsibility to runtime (through the `deploy.release_command` key in fly.toml). This required a change to use "cli-server" as the base image, since we need the "exo" binary to run `exo schema migrate`.

Also:
- Update Fly documentation to match the new implementation.
- Use 'fly' instead of 'flyctl' for brevity