refactor: Simplify ColumnTypeSpec
Earlier `ColumnTypeSpec` duplicated `PhysicalColumnType` except for an extra variation to define a foreign reference type.

Now `ColumnTypeSpec` holds two variants: `Direct(PhysicalColumnType)` and `Reference(ColumnReferenceSpec)`.

This is a preperatory step to make `PhysicalColumnType` extensible. Without this change, we would have to take care of of `ColumnTypeSpec` too in the same manner.