Refactor cast providers to make them extensible
Like the earlier refactoring, this makes introducing a new primitive type easier.

Instead of having a single `cast_value` function, we now have a registry of `CastProvider`s,
each of which is responsible for casting a specific type to a specific type.

Also fix a database solver test that was incorrect (compared an integer to a boolean) which
this refactoring uncovered.