Refactor cast providers to make them extensible (#1681)
Like the earlier refactorings, this makes introducing a new primitive type easier.

Instead of having a single `cast_value` function, we now have a registry of `CastProvider`s,
each of which is responsible for casting a specific value type to a specific SQL type.

Also fix a database solver test that was incorrect (compared an integer to a boolean) which
this refactoring uncovered.