Allow default values for decimal fields (#1695)
This pull request introduces the capability to define default values for `Decimal` fields using string literals in the schema (so that the precision is retained). It also enhances the database insertion logic to correctly handle these default values, including scenarios where all fields of a type have defaults and no explicit data is provided.

This enables models such as:

```
    type Event {
        @pk id: Int = autoIncrement()
        price: Decimal = "10.00"
    }
```