Add default value support for more types (#1696)
Following the default value support for `Decimal` in #1695, this change supports the same for `LocalDate`, `LocalTime`, `LocalDateTime`, `Json`, and `Uuid`.

We do not support default values for `Blob` (expect using a Postgres function). There is no semantically meaningful (and explainable) way to support it through a string and default for blob is uncommon.