feat(mcp): make the server's version compatible with the client's (#1713)
The [spec](https://modelcontextprotocol.io/specification/2025-03-26/basic/lifecycle#version-negotiation) requires that if we support the client's version, we return that to be the server's version. Since we support every version listed in `ProtocolVersion`, we return the client's version if it is valid.

This makes Exograph's MCP support compatible with the current Claude Desktop (which is still at "2024-11-05"), while supporting newer clients with support for the HTTP protocol.