fix(migration): improve column matching for constraints and indices
Match with name or column set associated with a constraint or an index. This removes migration's sensitivity to the names we generate vs names that a database may already have.

Without this change, we drop an index/constraint (with the old name) and create a new one with the same set of columns.