fix: deal with a race condition in  downloading Deno
When running integration tests for the first time when a new Deno version is introduced (in practice, this means once a local installation, but for every CI build), each test with a `@deno` module will try to install Deno and may run into a race condition.

This change makes use a file lock to co-ordinate.