refactor(deno): Use `deno bundle` (#1761)
Earlier, we did our own implementation to process source code (js/ts) to create Deno modules using a fork of Deno cli to create a library distribution (https://github.com/exograph/deno). Maintaining this fork required extensive effort.

With this PR, we use `deno bundle` that was reintroduced in version 2.4.x.

This is similar to what we used to do early on in Exograph. However, we make use experience simpler by automatically downloading the right version of Deno.